consolidates processing.imu_to_xyz into imu.to_xyz
